move game logic into command execution and model
starts game and waits for players to join
deals hands appropriately
collects button interactions for 'hit' and waits before disabling
ends session by showing dealer's face down card
still needs hand valuation and payout
