+++ app/events/ready.js
@@ -1,6 +1,5 @@
 const { User, Blackjack, sequelize, Op } = require('../db/models');
 const { Collection, OAuth2Guild } = require('discord.js');
-const { BlackjackGame } = require('../games/blackjack');
 
 module.exports = {
     name: 'ready',
@@ -33,7 +32,6 @@ module.exports = {
         // }));
 
         client.user.setPresence({ activities: [{ name: 'Keeping score', type: 'CUSTOM' }]});
-        await Blackjack.afterCreate(async (table, options) => new BlackjackGame(table, client))
         console.log(`Ready! Logged in as ${client.user.tag}`);
     },
 };
